Extract VideoDialogs component and consolidate all dialog refs into one
